pyproject.toml
[tool.poetry]
name = "verificacoes-github-splor-mg"
version = "0.1.0"
description = "A tool to fetch and iterate over GitHub organization repositories and issues."
authors = ["gabrielbdornas <gabrielbdornas@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
python-dotenv = "^1.0"
ipdb = "^0.13.13"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
